feat(main/elements): separate elements routes from elements module
