+++ src/app/main/elements/elements.module.ts
@@ -1,23 +1,10 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-const routes: Routes = [
-    {
-        path: '',
-        redirectTo: 'basic-ui',
-        pathMatch: 'full',
-    },
-    {
-        path: 'basic-ui',
-        loadChildren: './basic-ui/basic-ui.module#BasicUiModule',
-        data: { title: '基础 UI' }
-    }
-];
+import { ElementsRoutingModule } from './elements-routing.module';
 
 @NgModule({
-  imports: [
-      RouterModule.forChild(routes)
-  ],
-  declarations: []
+    imports: [
+        ElementsRoutingModule
+    ],
+    declarations: []
 })
 export class ElementsModule { }
